+++ src/main/java/com/example/demo/service/trust_score/impl/TrustScoreHistoryServiceImpl.java
@@ -1,6 +1,7 @@
-package com.example.demo.service.trust_score;
+package com.example.demo.service.trust_score.impl;
 
 import com.example.demo.repository.trust_score.TrustScoreHistoryRepository;
+import com.example.demo.service.trust_score.TrustScoreHistoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
