+++ src/main/java/com/example/demo/service/trust_score/impl/TrustScoreTypeServiceImpl.java
@@ -1,6 +1,7 @@
-package com.example.demo.service.trust_score;
+package com.example.demo.service.trust_score.impl;
 
 import com.example.demo.repository.trust_score.TrustScoreHistoryRepository;
+import com.example.demo.service.trust_score.TrustScoreTypeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
